refactor: extract feedResultToSchemaStore to shared utility
Moved feedResultToSchemaStore(), deserializeResultForSchema(), and
randomSample() from executePlaygroundCode.ts into a shared module at
src/documentdb/feedResultToSchemaStore.ts.

The playground now imports from the shared utility. The shell PTY will
use the same functions in the next commit.

No behavioral changes — pure extraction refactor.
